move methods to AssetGraphComputation
branch-name: mv-methods-to-asset-graph-computation
